refactor: reorganize dto and repository packages for improved structure #39
