+++ src/main/java/com/deview/server/domain/interview/dto/study/InterviewCategoryItemDto.java
@@ -1,4 +1,4 @@
-package com.deview.server.domain.interview.dto;
+package com.deview.server.domain.interview.dto.study;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
