+++ src/main/java/com/deview/server/domain/interview/repository/chat/ChatReviewRepository.java
@@ -1,4 +1,4 @@
-package com.deview.server.domain.interview.repository.review;
+package com.deview.server.domain.interview.repository.chat;
 
 import com.deview.server.domain.interview.domain.chat.ChatReview;
 import org.springframework.data.jpa.repository.JpaRepository;
