+++ src/main/java/com/deview/server/domain/interview/service/InterviewChatService.java
@@ -3,8 +3,8 @@ package com.deview.server.domain.interview.service;
 import com.deview.server.domain.interview.domain.chat.InterviewSession;
 import com.deview.server.domain.interview.domain.chat.ChatReview;
 import com.deview.server.domain.interview.dto.chat.Message;
-import com.deview.server.domain.interview.dto.chat.fastapi_request.FastApiEvaluationRequest;
-import com.deview.server.domain.interview.dto.chat.fastapi_request.FastApiNextRequest;
+import com.deview.server.domain.interview.dto.chat.request.fastapi_request.FastApiEvaluationRequest;
+import com.deview.server.domain.interview.dto.chat.request.fastapi_request.FastApiNextRequest;
 import com.deview.server.domain.interview.dto.chat.request.InterviewEvaluationRequest;
 import com.deview.server.domain.interview.dto.chat.request.InterviewNextRequest;
 import com.deview.server.domain.interview.dto.chat.request.InterviewStartRequest;
@@ -12,7 +12,7 @@ import com.deview.server.domain.interview.dto.chat.response.InterviewEvaluationR
 import com.deview.server.domain.interview.dto.chat.response.InterviewNextResponse;
 import com.deview.server.domain.interview.dto.chat.response.InterviewStartResponse;
 import com.deview.server.domain.interview.repository.chat.InterviewSessionRepository;
-import com.deview.server.domain.interview.repository.review.ChatReviewRepository;
+import com.deview.server.domain.interview.repository.chat.ChatReviewRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
